+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### _Feature_
 
 - Added `Sentence.getMetadata` to get the metadata of a sentence
+- Added dialog avatar APIs for character avatars, portrait-based avatar resolution, and the default dialog avatar display
 
 ## [0.9.0]